align with jbang-npm
